events/nats: setup consumer with FilterSubjects and rework PullMsg()
FilterSubjects() enables pull subscribers to filter
on multiple non-overlapping subjects.

Instead of requesting a batch of messages the PullOneMsg() method
will fetch one message at a time, this enables callers
to decide how messages are to be retrieved and errors
handled on a per subject basis instead of the previous
PullMsg() method where it would bail on the first error.